Decouple SimpleStorage from AWS
