+++ universal-application-tool-0.0.1/app/modules/CloudStorageModule.java
@@ -11,6 +11,8 @@ import controllers.dev.CloudEmulatorFileStorageStrategy;
 import play.Environment;
 import services.cloud.StorageClient;
 import services.cloud.StorageServiceName;
+import views.AwsFileUploadStrategy;
+import views.FileUploadStrategy;
 import views.dev.AwsStorageDevViewStrategy;
 import views.dev.AzureStorageDevViewStrategy;
 import views.dev.CloudStorageDevViewStrategy;
@@ -52,6 +54,9 @@ public class CloudStorageModule extends AbstractModule {
   }
 
   private void bindCloudStorageStrategy(String storageProvider) {
+    // Defaulting to this for now because prod file upload to Azure blob storage hasn't been
+    // implemented yet.
+    bind(FileUploadStrategy.class).to(AwsFileUploadStrategy.class);
     StorageServiceName storageServiceName = StorageServiceName.forString(storageProvider).get();
 
     switch (storageServiceName) {
