+++ universal-application-tool-0.0.1/app/views/AwsFileUploadStrategy.java
@@ -1,0 +1,124 @@
+package views;
+
+import static j2html.TagCreator.div;
+import static j2html.TagCreator.each;
+import static j2html.TagCreator.form;
+import static j2html.TagCreator.input;
+import static j2html.attributes.Attr.ENCTYPE;
+
+import controllers.applicant.routes;
+import j2html.attributes.Attr;
+import j2html.tags.ContainerTag;
+import j2html.tags.Tag;
+import java.util.Optional;
+import play.i18n.Messages;
+import play.mvc.Http.HttpVerbs;
+import services.applicant.question.FileUploadQuestion;
+import services.cloud.StorageUploadRequest;
+import services.cloud.aws.SignedS3UploadRequest;
+import views.applicant.ApplicantProgramBlockEditView.Params;
+import views.questiontypes.ApplicantQuestionRendererFactory;
+import views.questiontypes.ApplicantQuestionRendererParams;
+import views.style.BaseStyles;
+import views.style.ReferenceClasses;
+import views.style.Styles;
+
+public class AwsFileUploadStrategy extends FileUploadStrategy {
+
+  @Override
+  public ContainerTag signedFileUploadFields(
+      ApplicantQuestionRendererParams params, FileUploadQuestion fileUploadQuestion) {
+    StorageUploadRequest genericRequest = params.signedFileUploadRequest().get();
+    SignedS3UploadRequest request = (SignedS3UploadRequest) genericRequest;
+    Optional<String> uploaded =
+        fileUploadQuestion.getFilename().map(f -> String.format("File uploaded: %s", f));
+    ContainerTag fieldsTag =
+        div()
+            .with(div().withText(uploaded.orElse("")))
+            .with(input().withType("hidden").withName("key").withValue(request.key()))
+            .with(
+                input()
+                    .withType("hidden")
+                    .withName("success_action_redirect")
+                    .withValue(request.successActionRedirect()))
+            .with(
+                input()
+                    .withType("hidden")
+                    .withName("X-Amz-Credential")
+                    .withValue(request.credential()));
+    if (!request.securityToken().isEmpty()) {
+      fieldsTag.with(
+          input()
+              .withType("hidden")
+              .withName("X-Amz-Security-Token")
+              .withValue(request.securityToken()));
+    }
+    return fieldsTag
+        .with(input().withType("hidden").withName("X-Amz-Algorithm").withValue(request.algorithm()))
+        .with(input().withType("hidden").withName("X-Amz-Date").withValue(request.date()))
+        .with(input().withType("hidden").withName("Policy").withValue(request.policy()))
+        .with(input().withType("hidden").withName("X-Amz-Signature").withValue(request.signature()))
+        .with(input().withType("file").withName("file").attr(Attr.ACCEPT, acceptFileTypes()))
+        .with(errorDiv(params.messages(), fileUploadQuestion));
+  }
+
+  @Override
+  public Tag renderFileUploadBlockSubmitForms(
+      Params params, ApplicantQuestionRendererFactory applicantQuestionRendererFactory) {
+    // Note: This key uniquely identifies the file to be uploaded by the applicant and will be
+    // persisted in DB. Other parts of the system rely on the format of the key, e.g. in
+    // FileController.java we check if a file can be accessed based on the key content, so be extra
+    // cautious if you want to change the format.
+    String key =
+        String.format(
+            "applicant-%d/program-%d/block-%s/${filename}",
+            params.applicantId(), params.programId(), params.block().getId());
+    String onSuccessRedirectUrl =
+        params.baseUrl()
+            + routes.ApplicantProgramBlocksController.updateFile(
+                    params.applicantId(),
+                    params.programId(),
+                    params.block().getId(),
+                    params.inReview())
+                .url();
+    StorageUploadRequest request =
+        params.storageClient().getSignedUploadRequest(key, onSuccessRedirectUrl);
+
+    if (!(request instanceof SignedS3UploadRequest)) {
+      throw new RuntimeException(
+          "Tried to upload a file to AWS S3 storage using incorrect request type");
+    }
+    SignedS3UploadRequest signedRequest = (SignedS3UploadRequest) request;
+    ApplicantQuestionRendererParams rendererParams =
+        ApplicantQuestionRendererParams.builder()
+            .setMessages(params.messages())
+            .setSignedFileUploadRequest(signedRequest)
+            .build();
+
+    Tag uploadForm =
+        form()
+            .withId(BLOCK_FORM_ID)
+            .attr(ENCTYPE, "multipart/form-data")
+            .withAction(signedRequest.actionLink())
+            .withMethod(HttpVerbs.POST)
+            .with(
+                each(
+                    params.block().getQuestions(),
+                    question ->
+                        renderQuestion(
+                            question, rendererParams, applicantQuestionRendererFactory)));
+    Tag skipForms = renderDeleteAndContinueFileUploadForms(params);
+    Tag buttons = renderFileUploadBottomNavButtons(params);
+    return div(uploadForm, skipForms, buttons);
+  }
+
+  private String acceptFileTypes() {
+    return IMAGES_AND_PDF;
+  }
+
+  private ContainerTag errorDiv(Messages messages, FileUploadQuestion fileUploadQuestion) {
+    return div(fileUploadQuestion.fileRequiredMessage().getMessage(messages))
+        .withClasses(
+            ReferenceClasses.FILEUPLOAD_ERROR, BaseStyles.FORM_ERROR_TEXT_BASE, Styles.HIDDEN);
+  }
+}
